Refactor BdioDocument to eliminate internal state that is not configuration related
This eliminates the hackish builder pattern applied to BdioDocument and it's subclasses, it also makes the code easier to follow since the "read" operation returns a reactive stream and the "write" operation returns a subscriber. Finally, it eliminates the split configuration that occurred in graph mapping where the external context was computed and used in different projects.